+++ modules/apps/document-library/document-library-preview-api/bnd.bnd
@@ -1,4 +1,6 @@
 Bundle-Name: Liferay Document Library Preview API
 Bundle-SymbolicName: com.liferay.document.library.preview.api
-Bundle-Version: 2.0.1
-Export-Package: com.liferay.document.library.preview.background.task
+Bundle-Version: 2.1.0
+Export-Package:\
+	com.liferay.document.library.preview.background.task,\
+	com.liferay.document.library.preview.processor
