+++ modules/apps/document-library/document-library-preview-api/src/main/resources/com/liferay/document/library/preview/processor/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
